[AND-87] 포트폴리오 추가/수정 화면 작업 중/recyclerview, test등 필요

+++ app/src/main/java/kr/co/soogong/master/ui/requirements/progress/ProgressFragment.kt
@@ -91,7 +91,7 @@ class ProgressFragment : BaseFragment<FragmentRequirementsProgressBinding>(
         viewModel.event.observe(viewLifecycleOwner, EventObserver { (event, value) ->
             when (event) {
                 ProgressViewModel.BADGE_UPDATE -> {
-                    (parentFragment as? RequirementsBadge)?.setProgressBadge(value)
+                    (parentFragment as? RequirementsBadge)?.setProgressBadge(value as Int)
                 }
             }
         })
